HIIT-016 refactor: 예외 패키지 이동에 따른 반영
- dao 아래에서 domain 아래로 변경
